fix: XYNE-61737 decode FlowJSON in automation trigger message
Automations triggered by FlowJSON alert messages passed the raw stored
content (<div data-flow-json=...>Flow JSON</div>) straight into the agent's
task, so agents (e.g. rcaAgent) received the literal 'Flow JSON' placeholder
instead of the alert text and had to call other tools just to read what
triggered them.

- Extract the FlowJSON text helpers into shared utils/flowJson.ts (they were
  private to the zero messages side-effect handler, which already imports the
  message-received trigger — importing back would create a cycle).
- message-received.trigger.ts now decodes content via
  getFlowJsonContentForNotification (null-safe: normal messages pass through).
- messages-handler.ts re-uses the shared helpers (re-exports
  getFlowJsonContentForNotification so the vespa mapper import is unchanged).
